Volcano Demo Scene Update
Made several tweaks to the Obstacle Spawner Pool Editor for the volcanic rocks. 

First, I adjusted the spawn offset to go further off the screen. Now when you run full speed there are equal amounts of rocks that spawn on either side of you.

Second, I fixed a bug with setting the scale in the Obstacle Spawner Pool Editor. It works correctly now. 

Third, I added a z axis option in the editor so anyone can specify where in the z axis they want objects to spawn. 

Fourth, I created a new action on the state machine to disable shooting when the player gets knocked back, dies, or squished. The shooting is enabled again at the start of the iframes.

Lastly, I fixed a bug where the player got stuck in the crouch state when hit by a rock while crouching.